Remove Configuration V1
- consulcatalog
- docker
- marathon
- mesos
- ecs
- rancher
