--- provider/docker/config_root.go
@@ -1,12 +1,0 @@
-package docker
-
-import (
-	"github.com/containous/traefik/types"
-)
-
-func (p *Provider) buildConfiguration(containersInspected []dockerData) *types.Configuration {
-	if p.TemplateVersion == 1 {
-		return p.buildConfigurationV1(containersInspected)
-	}
-	return p.buildConfigurationV2(containersInspected)
-}
