+++ provider/docker/config_segment_test.go
@@ -721,7 +721,7 @@ func TestSegmentBuildConfiguration(t *testing.T) {
 				dockerDataList = append(dockerDataList, dData)
 			}
 
-			actualConfig := provider.buildConfigurationV2(dockerDataList)
+			actualConfig := provider.buildConfiguration(dockerDataList)
 			require.NotNil(t, actualConfig, "actualConfig")
 
 			assert.EqualValues(t, test.expectedBackends, actualConfig.Backends)
