+++ provider/ecs/config_test.go
@@ -1144,82 +1144,6 @@ func TestGetPort(t *testing.T) {
 	}
 }
 
-func TestGetFuncStringValue(t *testing.T) {
-	testCases := []struct {
-		desc         string
-		expected     string
-		instanceInfo ecsInstance
-	}{
-		{
-			desc:         "Protocol label is not set should return a string equals to http",
-			expected:     "http",
-			instanceInfo: simpleEcsInstance(map[string]*string{}),
-		},
-		{
-			desc:     "Protocol label is set to http should return a string equals to http",
-			expected: "http",
-			instanceInfo: simpleEcsInstance(map[string]*string{
-				label.TraefikProtocol: aws.String("http"),
-			}),
-		},
-		{
-			desc:     "Protocol label is set to https should return a string equals to https",
-			expected: "https",
-			instanceInfo: simpleEcsInstance(map[string]*string{
-				label.TraefikProtocol: aws.String("https"),
-			}),
-		},
-	}
-
-	for _, test := range testCases {
-		test := test
-		t.Run(test.desc, func(t *testing.T) {
-			t.Parallel()
-
-			actual := getFuncStringValueV1(label.TraefikProtocol, label.DefaultProtocol)(test.instanceInfo)
-			assert.Equal(t, test.expected, actual)
-		})
-	}
-}
-
-func TestGetFuncSliceString(t *testing.T) {
-	testCases := []struct {
-		desc         string
-		expected     []string
-		instanceInfo ecsInstance
-	}{
-		{
-			desc:         "Frontend entrypoints label not set should return empty array",
-			expected:     nil,
-			instanceInfo: simpleEcsInstance(map[string]*string{}),
-		},
-		{
-			desc:     "Frontend entrypoints label set to http should return a string array of 1 element",
-			expected: []string{"http"},
-			instanceInfo: simpleEcsInstance(map[string]*string{
-				label.TraefikFrontendEntryPoints: aws.String("http"),
-			}),
-		},
-		{
-			desc:     "Frontend entrypoints label set to http,https should return a string array of 2 elements",
-			expected: []string{"http", "https"},
-			instanceInfo: simpleEcsInstance(map[string]*string{
-				label.TraefikFrontendEntryPoints: aws.String("http,https"),
-			}),
-		},
-	}
-
-	for _, test := range testCases {
-		test := test
-		t.Run(test.desc, func(t *testing.T) {
-			t.Parallel()
-
-			actual := getFuncSliceStringV1(label.TraefikFrontendEntryPoints)(test.instanceInfo)
-			assert.Equal(t, test.expected, actual)
-		})
-	}
-}
-
 func makeEcsInstance(containerDef *ecs.ContainerDefinition) ecsInstance {
 	container := &ecs.Container{
 		Name:            containerDef.Name,
