+++ provider/marathon/config.go
@@ -25,7 +25,7 @@ type appData struct {
 	LinkedApps    []*appData
 }
 
-func (p *Provider) buildConfigurationV2(applications *marathon.Applications) *types.Configuration {
+func (p *Provider) buildConfiguration(applications *marathon.Applications) *types.Configuration {
 	var MarathonFuncMap = template.FuncMap{
 		"getDomain":      label.GetFuncString(label.TraefikDomain, p.Domain), // see https://github.com/containous/traefik/pull/1693
 		"getSubDomain":   p.getSubDomain,                                     // see https://github.com/containous/traefik/pull/1693
