--- templates/ecs-v1.tmpl
@@ -1,44 +1,0 @@
-[backends]
-{{range $serviceName, $instances := .Services }}
-  [backends."backend-{{ $serviceName }}".loadBalancer]
-    method = "{{ getLoadBalancerMethod $instances }}"
-    sticky = {{ getLoadBalancerSticky $instances }}
-
-    {{if hasStickinessLabel $instances }}
-    [backends."backend-{{ $serviceName }}".loadBalancer.stickiness]
-      cookieName = "{{ getStickinessCookieName $instances }}"
-    {{end}}
-
-    {{ if hasHealthCheckLabels $instances }}
-    [backends."backend-{{ $serviceName }}".healthCheck]
-      path = "{{ getHealthCheckPath $instances }}"
-      interval = "{{ getHealthCheckInterval $instances }}"
-    {{end}}
-
-    {{range $index, $i := $instances }}
-    [backends."backend-{{ $serviceName }}".servers."server-{{ $i.Name }}{{ $i.ID }}"]
-      url = "{{ getProtocol $i }}://{{ getHost $i }}:{{ getPort $i }}"
-      weight = {{ getWeight $i }}
-    {{end}}
-{{end}}
-
-[frontends]
-{{range $serviceName, $instances := .Services}}
-{{range filterFrontends $instances }}
-  [frontends."frontend-{{ $serviceName }}"]
-    backend = "backend-{{ $serviceName }}"
-    passHostHeader = {{ getPassHostHeader . }}
-    priority = {{ getPriority . }}
-
-    entryPoints = [{{range  getEntryPoints . }}
-      "{{.}}",
-      {{end}}]
-
-    basicAuth = [{{range getBasicAuth . }}
-      "{{.}}",
-    {{end}}]
-
-    [frontends."frontend-{{ $serviceName }}".routes."route-frontend-{{ $serviceName }}"]
-      rule = "{{getFrontendRule .}}"
-{{end}}
-{{end}}
